requirements.txt
boto3==1.17.75
botocore==1.20.75
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
colorlog==5.0.1
colormap==1.0.3
colorthief==0.2.1
easydev==0.11.1
h11==0.12.0
httpcore==0.12.3
httpx==0.17.1
idna==2.10
jmespath==0.10.0
labypy==1.0.0
libtools==0.3.3
pexpect==4.8.0
Pillow==8.2.0
progressbar==2.5
ptyprocess==0.7.0
Pygments==2.9.0
pypng==0.0.20
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
rfc3986==1.5.0
s3transfer==0.4.2
six==1.16.0
sniffio==1.2.0
spotipy==2.18.0
spotlib==0.2.10
tekore==3.7.1
urllib3==1.26.4
